server/user_service/requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
confluent-kafka==2.4.0
Django==5.0.6
django-cors-headers==4.4.0
django-prometheus==2.3.1
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.7
pillow==10.3.0
prometheus_client==0.20.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
redis==5.0.7
requests==2.32.3
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
